Move deprecation warning to existing drivers
